refactor(workspace): Decompose PersonRepository and update progress tracking
Decompose monolithic Person repository implementations into individual modules and update project progress documentation.

This commit refactors the Person repository in the [`banking-db-postgres`](banking-db-postgres) crate by splitting the monolithic [`src/repository/person/person_repository_impl.rs`](banking-db-postgres/src/repository/person/person_repository_impl.rs) and [`src/repository/person/person_repository_batch_impl.rs`](banking-db-postgres/src/repository/person/person_repository_batch_impl.rs) files into single-responsibility modules under [`src/repository/person/person_repository/`](banking-db-postgres/src/repository/person/person_repository/). Corresponding tests have been co-located with their implementation files. Additionally, progress documentation has been updated in [`docs/progress/progress-tracking.md`](docs/progress/progress-tracking.md) to reflect the new modular structure for the Person repository.

Components affected:
- [`banking-db-postgres/src/repository/person/person_repository_impl.rs`](banking-db-postgres/src/repository/person/person_repository_impl.rs)
- [`banking-db-postgres/src/repository/person/person_repository_batch_impl.rs`](banking-db-postgres/src/repository/person/person_repository_batch_impl.rs)
- [`banking-db-postgres/src/repository/person/person_repository/`](banking-db-postgres/src/repository/person/person_repository/)
- [`docs/progress/progress-tracking.md`](docs/progress/progress-tracking.md)

Benefits:
- Improved modularity and maintainability through single-responsibility files
- Enhanced readability and test co-location
- Reduced file size for better performance with LLMs and tooling
- Accurate project progress tracking documentation
