+++ banking-db-postgres/src/lib.rs
@@ -9,7 +9,7 @@ pub use repository::person::country_subdivision_repository::CountrySubdivisionRe
 pub use repository::person::entity_reference_repository::EntityReferenceRepositoryImpl;
 pub use repository::person::locality_repository::LocalityRepositoryImpl;
 pub use repository::person::location_repository::LocationRepositoryImpl;
-pub use repository::person::person_repository_impl::PersonRepositoryImpl;
+pub use repository::person::person_repository::PersonRepositoryImpl;
 pub use repository::unit_of_work_impl;
 #[cfg(test)]
 pub mod test_helper;
