--- banking-db-postgres/tests/suites/person/helpers.rs
@@ -1,103 +1,0 @@
-use banking_db::models::person::{
-    CountryModel, CountrySubdivisionModel, EntityReferenceModel, LocalityModel, LocationModel,
-    LocationType, PersonModel, PersonType, RelationshipRole,
-};
-use heapless::String as HeaplessString;
-use uuid::Uuid;
-
-pub fn create_test_person_model(name: &str) -> PersonModel {
-    PersonModel {
-        id: Uuid::new_v4(),
-        person_type: PersonType::Natural,
-        display_name: HeaplessString::try_from(name).unwrap(),
-        external_identifier: Some(
-            HeaplessString::try_from(Uuid::new_v4().to_string().as_str()).unwrap(),
-        ),
-        entity_reference_count: 0,
-        organization_person_id: None,
-        messaging_info1: None,
-        messaging_info2: None,
-        messaging_info3: None,
-        messaging_info4: None,
-        messaging_info5: None,
-        department: None,
-        location_id: None,
-        duplicate_of_person_id: None,
-    }
-}
-
-pub fn create_test_country_model(iso2: &str, name_l1: &str) -> CountryModel {
-    CountryModel {
-        id: Uuid::new_v4(),
-        iso2: HeaplessString::try_from(iso2).unwrap(),
-        name_l1: HeaplessString::try_from(name_l1).unwrap(),
-        name_l2: None,
-        name_l3: None,
-    }
-}
-
-pub fn create_test_country_subdivision_model(
-    country_id: Uuid,
-    code: &str,
-    name_l1: &str,
-) -> CountrySubdivisionModel {
-    CountrySubdivisionModel {
-        id: Uuid::new_v4(),
-        country_id,
-        code: HeaplessString::try_from(code).unwrap(),
-        name_l1: HeaplessString::try_from(name_l1).unwrap(),
-        name_l2: None,
-        name_l3: None,
-    }
-}
-
-pub fn create_test_locality_model(
-    country_subdivision_id: Uuid,
-    code: &str,
-    name_l1: &str,
-) -> LocalityModel {
-    LocalityModel {
-        id: Uuid::new_v4(),
-        country_subdivision_id,
-        code: HeaplessString::try_from(code).unwrap(),
-        name_l1: HeaplessString::try_from(name_l1).unwrap(),
-        name_l2: None,
-        name_l3: None,
-    }
-}
-
-pub fn create_test_location_model(
-    locality_id: Uuid,
-    street_line1: &str,
-    postal_code: &str,
-) -> LocationModel {
-    LocationModel {
-        id: Uuid::new_v4(),
-        location_type: LocationType::Residential,
-        street_line1: HeaplessString::try_from(street_line1).unwrap(),
-        street_line2: None,
-        street_line3: None,
-        street_line4: None,
-        locality_id,
-        postal_code: Some(HeaplessString::try_from(postal_code).unwrap()),
-        latitude: None,
-        longitude: None,
-        accuracy_meters: None,
-    }
-}
-
-pub fn create_test_entity_reference_model(
-    person_id: Uuid,
-    entity_role: RelationshipRole,
-    reference_external_id: &str,
-) -> EntityReferenceModel {
-    EntityReferenceModel {
-        id: Uuid::new_v4(),
-        person_id,
-        entity_role,
-        reference_external_id: HeaplessString::try_from(reference_external_id).unwrap(),
-        reference_details_l1: None,
-        reference_details_l2: None,
-        reference_details_l3: None,
-    }
-}
