--- banking-db-postgres/tests/suites/person/mod.rs
@@ -1,5 +1,0 @@
-// pub mod location_repository_tests;
-pub mod person_repository_tests;
-pub mod person_batch_operations_test;
-//pub mod location_batch_operations_test;
-pub mod helpers;
